Removed the title monitoring, as it was dependent on to many factory (Edge cutting the title).
